server/requirements.txt
transformers 
pytorch